Resolve data type mismatch for plan_id in SQL schema
- Corrected a data type mismatch in `ShopCore.sql` where `plan_id` was incorrectly defined as `INT` instead of `SMALLINT`.
- Aligned `order_details` schema with the master `plans` table to ensure referential integrity and support the foreign key constraint.
- Updated `orderRepository.js` to use `sql.SmallInt` for plan-related database inputs, preventing potential type conversion errors.
- Finalized the integrated order-subscription database structure for mixed cart transactions.

+++ database/Querys/ShopCore.sql
@@ -50,7 +50,7 @@ CREATE TABLE [dbo].[order_details] (
 	[detail_id] BIGINT IDENTITY(1,1) PRIMARY KEY,
 	[order_id] BIGINT NOT NULL,
 	[product_id] INT NULL,
-	[plan_id] INT NULL,
+	[plan_id] SMALLINT NULL,
 	[quantity] SMALLINT NOT NULL, 
 	[unit_price] DECIMAL(10, 2) NOT NULL, 
 	FOREIGN KEY (order_id) REFERENCES orders(order_id),
